move get_multipole_matrix to its own module #98
