+++ nac/integrals/__init__.py
@@ -1,4 +1,5 @@
 from .fourierTransform import (calculate_fourier_trasform_cartesian)
+from .matrix_multipole import (compute_matrix_multipole, get_multipole_matrix)
 from .multipoleIntegrals import (calcMtxMultipoleP, general_multipole_matrix)
 from .nonAdiabaticCoupling import (calculate_couplings_levine,
                                    calculate_couplings_3points,
@@ -11,5 +12,5 @@ from .spherical_Cartesian_cgf import calc_transf_matrix
 __all__ = ['calculate_couplings_3points', 'calculate_couplings_levine',
            'calc_transf_matrix', 'calcMtxMultipoleP', 'calcMtxOverlapP',
            'calculate_couplings_levine', 'calculate_fourier_trasform_cartesian',
-           'compute_overlaps_for_coupling', 'correct_phases',
-           'general_multipole_matrix']
+           'compute_matrix_multipole', 'compute_overlaps_for_coupling', 'correct_phases',
+           'general_multipole_matrix', 'get_multipole_matrix']
